wip: CPM analytics hook
